Fix some stuff
Newline stripping was broken since the non-stripped version was discarded so there
was no way to realize we needed to upload the stripped version to correct the
trailing newline on VNT.

We print the number of updated translations to upload before confirming with the
user, and also we properly repeat the prompt if something other than "yes" or
"no" is entered.

The check for empty VNT scripts was looking for the line count to equal "0"
rather than 0.