Fix an outdated type signature
This wasn't caught since I didn't specify what type of Iterator the argument
was.  It would be nice if it was possible to catch this, but I guess Python's
type annotations aren't expressive enough to do so... so instead of worrying
about it I just made the function return a bool instead.